requirements.txt
asgiref==3.5.2
certifi==2022.6.15.1
charset-normalizer==2.1.1
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
idna==3.3
Pillow==9.2.0
PyJWT==2.4.0
python-decouple==3.6
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
twilio==7.14.0
urllib3==1.26.12
